move to plone5
So far I updated the test code so that real problems can be handled
In this case content is created with the assumption that it will be
archetypes objects. This fails here because images seem to need some
attribute. plone.app.contenttypes can be used to check for examples how
to create images.
I asked on irc if we need AT compatibility.
2 Answered with screw it, so go for it