Avoid unnecessarily changing Diazo theme files when making dev bundle
When making a development JS bundle, don't use "sed" to rename the JS bundle
being used in the Diazo theme files.

Otherwise we have a bunch of dirty *.html files that should never be committed,
causing a hazard.

Instead, we simply copy over the ploneintranet-dev.js file to
theme/static/generated/bundles/ploneintranet.js, which is the file being used
by the Diazo theme files and which is not under version control.