Visit send-message in browser to send message to admin.
This adds a Celery task.
This task sends a message to Tornado.
Tornado sends a message to a websocket.
The websocket should probably already have been opened.
Note that this uses Redis, which needs to run,
and Celery and Tornady which both need to run,
which is not really handled by the buildout yet.