plone.belowcontent: don't register PI.simplesharing.share
A Share link is rendered in plone.belowcontent via the main_template,
this appears on /workspaces/ although it doesn't make much sense to
share such a listing page and it's not in proto. There is a placeholder
for sharing documents in the metadata bar so I doubt we ever want to
include a share button in the page footer.