event_view: remove duplicate Description field / hide metadata
In the prototype there is a description field in both the body and the
collapsible "metadata" field. We can't submit two fields with the same
name (or we end up with a list of strings). Since there are no metadata
fields in the collapsible menu, I've commented it out altogether.