dexterity_update: reset value to default when we get NO_VALUE
I suspect this has just worked by fluke up to now, e.g. for tags, the
extract method returns an empty tuple instead of NO_VALUE, so the field
value is actually updated. We now need to use a vocabulary for a field
for iKath and resetting it doesn't update the field.