Search results: translate friendly_type_name
We index the mime type name for files e.g. "Word 2007 document" and
for plone content types we use the name of the content type. Although
Products.Mimetypes has a registry where we can look up the relevant
translation for each file type it's tricky to do that for the search
facets since we don't index the id.

As a work-around, this uses the ploneintranet i18n domain and adds
msgids for the various content types and file types.

(cherry picked from commit 13c2f78c67d4946bdb85ece4d5214cafa38bcbf6)