update_todos_state: perform transition as temporary Editor
This allows Todo transitions to be protected by the Editor role, and
also for the workflow state to be set to Open or Planned as
appropriate.

Without this change it is possible for a user to perform a transition on
a Case which puts it into a state where that user no longer has the
permissions to edit Todo items. The event handler will then try to
update the workflow state of Todo items e.g. from Planned to Open and
this will fail, which will cause the whole transaction to fail.