member_prefill: don't restrict to workspace members
The prefill only needs to return the details for the users selected in
the field. It doesn't need to check whether the users are members of the
workspace or not. The available users for a field are determined by the
ajax-url.

(cherry picked from commit a172d92e09a63c91363553d3d4f4f85565cd155f)