Create Workspace: avoid acl_users caching issues
When caching is enabled on acl_users, creating a new workspace fails.

This bypasses security checks for transitioning a freshly created
workspace to the correct state and then invalidates the cache so that
the user who created the workspace has the appropriate roles on the
workspace.

NOTE: this certainly needs cleanup, but it should be enough to allow us
to cache acl_users.

This modifies ploneintranet_workspace_workflow so that the workspace
Owner has "Access contents information" in the default workflow
state. To test, you will need to import the workflow, or apply the
change manually.