Fix dashboard workspace tile headings
In e3078070 the logic for the workspace tiles was refactored, but the
workspace_type parameter was also being used for the tile heading so the
headings were no longer translated.

I re-ran the synci18n.sh script to update the translations which also
includes some extra changes:

The following msgids were removed:
    #: ../userprofile/browser/templates/user_import_form.pt:13
    msgid "Bulk import users"
    msgstr ""

    #: ../workspace/browser/tiles/templates/sidebar.pt:417
    msgid "Starts"
    msgstr ""

    #: ../workspace/browser/tiles/templates/sidebar.pt:422
    msgid "Ended"
    msgstr ""

This is correct, since the sidebar was updated in 9c41141da and
user_import_form was updated in 3544add4.

(cherry picked from commit 789406c437910d2b161e7fca050d1659116a6dcc)