case_workflow_guard: only consider open Tasks for the current milestone
It is currently possible to have an open Task assigned to a closed
milestone in a number of ways:

  1. unchecking a Task in the sidebar
  2. changing the state in the edit view
  3. creating a new Task assigned to a closed milestone
  4. assigning an existing Task to a closed milestone

In the MetroMap an open Task in a closed milestone has the class
'state-open' but it is not displayed because it is within an <li> which
has the class 'state-finished'. This means you can see that a Task is
open in the sidebar, but not in the MetroMap.

Prior to this change, an open Task in a closed milestone would also
prevent someone from closing the current milestone, which is quite
confusing.

Perhaps we should prevent open tasks ever being assigned to closed
milestones in the cases listed above.

(cherry picked from commit 3cf74ccb97a789c733beddd0f74b535b733a59b5)