Set default body text for Documents
Previously we injected content into the body of a Document if it was
empty so that a user has an area available in the editor where they can
click to start editing. This triggered a change to the document which
showed a confirmation dialog when a user navigated away from the
Document, (because of unsaved changes).

This change sets the default text on creation to avoid this issue.