Division fields in the workspace sidebar
Show the is_division field in the advanced sidebar only
if it has meaning for the context.

Disable the is_division field for Cases

In the advanced setting tab of the workspace sidebar
it was possible to uncheck the is_division input
and select another division even if we were not authorized.

This patch fixes this UI inconsistency.

Server side the view was already properly secured.