Bulk cut: notify if selected objs can't be cut
A selected item may be locked, or the user may not have permission to
delete it so a notification is now shown for items which can't be cut.

Also, since an item from another workspace could be in the shopping cart
the cart-cut view isn't protected by the ModifyPortalContent permission.