Save the archival_date when checked in the sidebar
This requires a custom widget so that the extract method returns None
when the field is unchecked, which sets the field to None.

I'm not sure it's a good idea to conflate storing the date when the
field is checked and the fact that it is checked. We need to use the
date in the sidebar (the object context) and we only need to be able to
search by is_archived, so I've added an boolean indexer for that.

Alternatively, we could add an extra field to store the date and update
that using an event handler. That seems like overkill to me, but perhaps
it's more elegant.