Do not limit search results we are looking for the section children
By default the query that was in place, returned just the first 10 results.

This patch sets the limit to a very high number.

A test is ready, but the test suite uses zcatalog, while the library utils rely heavily on solr, so the tested has to be executed within a context manager with the promise that one day either the utils will use a way to get the results which is independent on the search backend, either we have solr in pla ce in the testing suite.