Synchronous previews when creating testing content
At the moment we have a conflict error when generating testing content
with an instance in foreground.

I identified the culprit in the generation of the previews, altough I
still do not know who is conflicting with it.

This PR decorates the setuphandler with a method that disables
temporarily the asynchronous preview generation and fixes the race
condition.

Fixes #280