Save PDF versions of Documents for bulk download (#655)
* Generate PDF versions of Documents

This is mostly copied from Star, except that it uses
collective.documentviewer to generate the PDF files.

* Add @@pdf view for Docs and include PDF versions in the bulk download

* PDF: Factor out the list of HTML_CONTENTTYPES

* Test the PDF view

* HTMLConv: standardize method names

* PDF: ensure image paths are traverseable from root

* Add tests for HTML to PDF conversion

* PDF: generate for empty docs

c.documentviewer expects certain folders to be created e.g. the large
folder for preview images. Since having Document objects with no content
is an edge case anyway, let's generate PDFs for them to be consistent.

* Bulk download: fix warnings for undownloadable items

* Bulk download add robot test for undownloadables

* tidy up html_converter

* test_pdf_view: PDFs are generated automatically

We don't need to explicitly call generate_pdf, and we now also generate
PDFs for empty documents, to be compatible with the way c.documentviewer
handles files.

* Code style fix

* docconv.test_views remove unused import