avatar import for profile names containing dots
as plone normalizes john.doe.jpg to john_doe.jpg the @@import-avatars view needs to handle that special case.
(this fixes #1117)

note that we got rid of `transaction.commit()` in the import view since this broke test-isolation.
the only drawback i see here is that no avatar image gets imported if the view raises an error.